bot/requirements.txt
aiogram==3.0.0
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==0.21.0
